bytecler/tgface/requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0
